FORGE-890 Implementation of the new Scaffold-X plugin and API.
Contains a collection of new APIs for implementing Scaffold-X providers (that support scaffolding from X resource). This includes a Freemarker API that can be used by providers, to load templates and provide generated code. This also includes a Metawidget based API for inspecting Forge JavaClasses and obtaining inspection results.

Contains a delegating scaffold provider for Faces, so that the existing Faces provider of Forge is reused. Tests for this provider are distinct from the tests in the existing provider.