FORGE-913 Added support for allowing timestamp input via Richfaces calendar
This adds support for supplying timestamps via the generated scaffold, with a minor disadvantage - the time is defaulted to 12:00PM, but can be modified again. Fixing this requires adding support in Metawidget to display the 'apply' button.

Additionally, this still does not support the input of time fields, since Richfaces calendar does not support this component yet.