FORGE-949 Corrected validation plugin to emit descriptors with only default provider values.
MessageInterpolator, TraversableResolver and ConstraintValidatorFactory values are set only when provided by the user. This will ensure that the default values as supplied by the provider will be used unless the user overrides them. This would prevent issues involving changes in the internals of the BV providers, as was the case with HV-484.